Move types to shared file
