+++ packages/nouns-contracts/tasks/deploy-local.ts
@@ -2,19 +2,9 @@ import { default as NounsAuctionHouseABI } from '../abi/contracts/NounsAuctionHo
 import { task, types } from 'hardhat/config';
 import { Interface } from 'ethers/lib/utils';
 import { Contract as EthersContract } from 'ethers';
+import { ContractName } from './types';
 
-type ContractName =
-  | 'WETH'
-  | 'NFTDescriptor'
-  | 'NounsDescriptor'
-  | 'NounsSeeder'
-  | 'NounsToken'
-  | 'NounsAuctionHouse'
-  | 'NounsAuctionHouseProxyAdmin'
-  | 'NounsAuctionHouseProxy'
-  | 'NounsDAOExecutor'
-  | 'NounsDAOLogicV1'
-  | 'NounsDAOProxy';
+type LocalContractName = ContractName | 'WETH';
 
 interface Contract {
   args?: (string | number | (() => string | undefined))[];
@@ -61,12 +51,12 @@ task('deploy-local', 'Deploy contracts to hardhat')
       from: deployer.address,
       nonce: nonce + AUCTION_HOUSE_PROXY_NONCE_OFFSET,
     });
-    const contracts: Record<ContractName, Contract> = {
+    const contracts: Record<LocalContractName, Contract> = {
       WETH: {},
       NFTDescriptor: {},
       NounsDescriptor: {
         libraries: () => ({
-          NFTDescriptor: contracts['NFTDescriptor'].instance?.address as string,
+          NFTDescriptor: contracts.NFTDescriptor.instance?.address as string,
         }),
       },
       NounsSeeder: {},
@@ -74,8 +64,8 @@ task('deploy-local', 'Deploy contracts to hardhat')
         args: [
           args.noundersdao || deployer.address,
           expectedAuctionHouseProxyAddress,
-          () => contracts['NounsDescriptor'].instance?.address,
-          () => contracts['NounsSeeder'].instance?.address,
+          () => contracts.NounsDescriptor.instance?.address,
+          () => contracts.NounsSeeder.instance?.address,
           proxyRegistryAddress,
         ],
       },
@@ -85,12 +75,12 @@ task('deploy-local', 'Deploy contracts to hardhat')
       NounsAuctionHouseProxyAdmin: {},
       NounsAuctionHouseProxy: {
         args: [
-          () => contracts['NounsAuctionHouse'].instance?.address,
-          () => contracts['NounsAuctionHouseProxyAdmin'].instance?.address,
+          () => contracts.NounsAuctionHouse.instance?.address,
+          () => contracts.NounsAuctionHouseProxyAdmin.instance?.address,
           () =>
             new Interface(NounsAuctionHouseABI).encodeFunctionData('initialize', [
-              contracts['NounsToken'].instance?.address,
-              contracts['WETH'].instance?.address,
+              contracts.NounsToken.instance?.address,
+              contracts.WETH.instance?.address,
               args.auctionTimeBuffer,
               args.auctionReservePrice,
               args.auctionMinIncrementBidPercentage,
@@ -106,11 +96,11 @@ task('deploy-local', 'Deploy contracts to hardhat')
       },
       NounsDAOProxy: {
         args: [
-          () => contracts['NounsDAOExecutor'].instance?.address,
-          () => contracts['NounsToken'].instance?.address,
+          () => contracts.NounsDAOExecutor.instance?.address,
+          () => contracts.NounsToken.instance?.address,
           args.noundersdao || deployer.address,
-          () => contracts['NounsDAOExecutor'].instance?.address,
-          () => contracts['NounsDAOLogicV1'].instance?.address,
+          () => contracts.NounsDAOExecutor.instance?.address,
+          () => contracts.NounsDAOLogicV1.instance?.address,
           args.votingPeriod,
           args.votingDelay,
           args.proposalThresholdBps,
